+++ src/main/java/com/b1n_ry/yigd/block/entity/GraveBlockEntity.java
@@ -47,14 +47,10 @@ public class GraveBlockEntity extends BlockEntity {
     private boolean glowing;
 
     public GraveBlockEntity(BlockPos pos, BlockState state) {
-        this(null, pos, state);
-    }
-    public GraveBlockEntity(String customName, BlockPos pos, BlockState state) {
         super(Yigd.GRAVE_BLOCK_ENTITY, pos, state);
 
         this.graveOwner = null;
         this.storedXp = 0;
-        this.customName = customName;
         this.storedInventory = DefaultedList.ofSize(41, ItemStack.EMPTY);
 
         this.creationTime = world != null ? world.getTime() : 0;
@@ -160,6 +156,8 @@ public class GraveBlockEntity extends BlockEntity {
 
         // Not from NBT. Static definitions
         this.glowing = YigdConfig.getConfig().graveSettings.graveRenderSettings.glowingGrave;
+
+        this.markDirty();
     }
 
     @Override