+++ src/main/java/com/b1n_ry/yigd/core/ServerPacketReceivers.java
@@ -77,6 +77,7 @@ public class ServerPacketReceivers {
             } else if (!DeathInfoManager.INSTANCE.unlockedGraves.contains(graveId)) {
                 DeathInfoManager.INSTANCE.unlockedGraves.add(graveId);
             }
+            DeathInfoManager.INSTANCE.markDirty();
         });
     }
 }